Somewhat finer grained process control
svn path=/trunk/KDE/kdebase/runtime/; revision=794047

+++ nepomuk/server/processcontrol.cpp
@@ -61,8 +61,9 @@ void ProcessControl::setCrashPolicy( CrashPolicy policy )
 void ProcessControl::stop()
 {
     if ( mProcess.state() != QProcess::NotRunning ) {
-        mProcess.waitForFinished( 30000 );
-        mProcess.terminate();
+        if ( !mProcess.waitForFinished( 30000 ) ) {
+            mProcess.terminate();
+        }
     }
 }
